schema rework for correct values
