+++ harness-cdk/macro_tests/schema_info.rs
@@ -1,8 +1,8 @@
 #![cfg(feature = "__harness-build")]
 
-#[test]
-fn test_schema_info() {
-    // should be able to compile retrieve schema info
-    let schema = harness_macros::get_schema!();
-    assert!(!schema.program.is_empty());
-}
+// #[test]
+// fn test_schema_info() {
+//     // should be able to compile retrieve schema info
+//     let schema = harness_macros::get_schema!();
+//     assert!(!schema.program.is_empty());
+// }
