+++ harness-primitives/src/internals.rs
@@ -7,27 +7,11 @@ use serde::{Deserialize, Serialize};
 /// consumer, but rather by the `harness` macro to generate the necessary code.
 ///
 /// Ok to access once it's present in the [`Program`](crate::program::Program) struct.
-#[derive(Deserialize, Serialize, Clone, Debug, CandidType)]
+#[derive(Deserialize, Default, Serialize, Clone, Debug, CandidType)]
 pub struct Schema {
     pub version: String,
     pub program: String,
     pub services: Vec<Service>,
-}
-
-impl Default for Schema {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl Schema {
-    pub fn new() -> Self {
-        Self {
-            version: "-".to_string(),
-            program: "-".to_string(),
-            services: vec![],
-        }
-    }
 }
 
 /// This is the intermediate schema that is used to generate the code for the harness program. It
